backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Security Scanners
bandit==1.7.5
semgrep==1.45.0
pip-audit==2.6.1

# ML/AI
transformers==4.35.0
torch==2.1.0
sentence-transformers==2.2.2

# GitHub Integration
PyGithub==2.1.1
GitPython==3.1.40

# Database
pymongo==4.6.0
dnspython==2.4.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
pandas==2.1.3
numpy==1.26.2
reportlab==4.0.7
PyJWT==2.8.0
clerk-backend-api==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
